+++ src/simp/simp.h
@@ -1,4 +1,4 @@
-// Vendored from gcol33/SIMP (a87642d) by vendor_simp.sh -- do not edit.
+// Vendored from gcol33/SIMP (cc1de1d) by vendor_simp.sh -- do not edit.
 // Upstream: SIMP inst/include/simp/simp.h. Edit there and re-vendor.
 
 #ifndef SIMP_SIMP_H
@@ -8,6 +8,9 @@
 // with LinkingTo: SIMP and include <simp/simp.h>.
 #include "scheme.h"
 #include "integrate.h"
+#include "mts.h"
 #include "implicit.h"
+#include "harmonic.h"
+#include "adapt.h"
 
 #endif  // SIMP_SIMP_H